reorganize region files
