+++ src/common/regions/region_hooks.ts
@@ -1,7 +1,13 @@
-import { useContext } from 'react';
+import { createContext, useContext } from 'react';
 import { useParams } from 'react-router';
-import regions, { Region, RegionContext } from 'common/regions';
 import { assert } from 'common/utils';
+import regions from './region_db';
+import { Region } from './types';
+
+// We are careful to never call `useRegion()` in components that are not
+// nested inside `<RegionContext.Provider>` so we cheat and pretend that the
+// value is non-nullable
+export const RegionContext = createContext<Region>(null as any);
 
 export const useLocationPageRegion = () => {
   const region = useContext(RegionContext);
