+++ src/common/regions/types.ts
@@ -1,0 +1,91 @@
+import urlJoin from 'url-join';
+import { getAbbreviatedCounty } from 'common/utils/compare';
+
+export type FipsCode = string;
+
+export enum RegionType {
+  COUNTY = 'county',
+  STATE = 'state',
+  CBSA = 'CBSA',
+}
+
+export abstract class Region {
+  constructor(
+    public readonly name: string,
+    public readonly urlSegment: string,
+    public readonly fipsCode: FipsCode,
+    public readonly population: number,
+    public readonly regionType: RegionType,
+  ) {}
+
+  abstract get fullName(): string;
+  abstract get shortName(): string;
+  abstract get abbreviation(): string;
+  abstract get relativeUrl(): string;
+
+  get canonicalUrl() {
+    return urlJoin('https://covidactnow.org', this.relativeUrl);
+  }
+}
+
+export class State extends Region {
+  constructor(
+    name: string,
+    urlSegment: string,
+    fipsCode: FipsCode,
+    population: number,
+    public readonly stateCode: string,
+  ) {
+    super(name, urlSegment, fipsCode, population, RegionType.STATE);
+  }
+
+  get fullName() {
+    return this.name;
+  }
+
+  get shortName() {
+    return this.name;
+  }
+
+  get abbreviation() {
+    return this.stateCode;
+  }
+
+  get relativeUrl() {
+    return `us/${this.urlSegment}`;
+  }
+}
+
+export class County extends Region {
+  constructor(
+    name: string,
+    urlSegment: string,
+    fipsCode: FipsCode,
+    population: number,
+    public readonly state: State,
+    public readonly cityNames: string[],
+    private readonly adjacentCountiesFips: FipsCode[],
+  ) {
+    super(name, urlSegment, fipsCode, population, RegionType.COUNTY);
+  }
+
+  get fullName() {
+    return `${this.name}, ${this.state.name}`;
+  }
+
+  get shortName() {
+    return `${this.name}, ${this.stateCode}`;
+  }
+
+  get relativeUrl() {
+    return urlJoin(this.state.relativeUrl, `county/${this.urlSegment}`);
+  }
+
+  get abbreviation() {
+    return getAbbreviatedCounty(this.name);
+  }
+
+  get stateCode() {
+    return this.state.stateCode;
+  }
+}
